// apps/react-frontend/src/api/client/models/transport-view.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ProfileDto } from './profile-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { Truck } from './truck';

/**
 * 
 * @export
 * @interface TransportView
 */
export interface TransportView {
    /**
     * 
     * @type {string}
     * @memberof TransportView
     */
    'pickupDate': string;
    /**
     * 
     * @type {string}
     * @memberof TransportView
     */
    'deliveryDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof TransportView
     */
    'id': string;
    /**
     * 
     * @type {Truck}
     * @memberof TransportView
     */
    'truck'?: Truck;
    /**
     * 
     * @type {string}
     * @memberof TransportView
     */
    'originCity'?: string;
    /**
     * 
     * @type {string}
     * @memberof TransportView
     */
    'destinationCity'?: string;
    /**
     * 
     * @type {ProfileDto}
     * @memberof TransportView
     */
    'driver'?: ProfileDto;
    /**
     * 
     * @type {string}
     * @memberof TransportView
     */
    'status': TransportViewStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof TransportView
     */
    'displayNumber'?: string;
    /**
     * 
     * @type {string}
     * @memberof TransportView
     */
    'containerId'?: string;
    /**
     * 
     * @type {string}
     * @memberof TransportView
     */
    'transportType': TransportViewTransportTypeEnum;
    /**
     * 
     * @type {number}
     * @memberof TransportView
     */
    'sequenceNumber': number;
}

export const TransportViewStatusEnum = {
    Unplanned: 'UNPLANNED',
    Planned: 'PLANNED',
    Finished: 'FINISHED',
    Invoiced: 'INVOICED'
} as const;

export type TransportViewStatusEnum = typeof TransportViewStatusEnum[keyof typeof TransportViewStatusEnum];
export const TransportViewTransportTypeEnum = {
    Container: 'CONTAINER',
    Waste: 'WASTE'
} as const;

export type TransportViewTransportTypeEnum = typeof TransportViewTransportTypeEnum[keyof typeof TransportViewTransportTypeEnum];